Simplify the setup
