+++ .gitignore
@@ -27,9 +27,7 @@ Carthage/Build
 .swiftpm/xcode/xcuserdata
 .swiftpm/package.xcworkspace/xcuserdata
 ### Samples ###
-/Derived
-/KSCrash.xcodeproj
-/KSCrashSampleWorkspace.xcworkspace
 /Samples/Common/.swiftpm
 /Samples/**/Derived
 /Samples/**/*.xcodeproj
+/Samples/**/*.xcworkspace
