+++ Samples/Sources/App.swift
@@ -29,6 +29,10 @@ import SampleUI
 
 @main
 struct SampleApp: App {
+#if os(macOS)
+    @NSApplicationDelegateAdaptor(AppDelegate.self) var appDelegate
+#endif
+
     var body: some Scene {
         WindowGroup {
             SampleView()
