--- Samples/iOS/Project.swift
@@ -1,28 +1,0 @@
-import ProjectDescription
-
-let project = Project(
-    name: "iOS",
-    packages: [
-        .local(path: "../Common"),
-    ],
-    targets: [
-        .target(
-            name: "iOS",
-            destinations: .iOS,
-            product: .app,
-            bundleId: "com.github.kstenerud.KSCrash.iOS",
-            infoPlist: InfoPlist.extendingDefault(with: [
-                "UILaunchScreen": [
-                    "UIImageName": "LaunchImage",
-                    "UIBackgroundColor": "LaunchScreenColor",
-                ],
-                "UISupportedInterfaceOrientations": ["UIInterfaceOrientationPortrait"],
-                "CFBundleDisplayName": "KSCrash iOS",
-            ]),
-            sources: ["Sources/**"],
-            dependencies: [
-                .package(product: "SampleUI", type: .runtime),
-            ]
-        ),
-    ]
-)
