--- Workspace.swift
@@ -1,12 +1,0 @@
-import ProjectDescription
-
-let workspace = Workspace(
-    name: "KSCrashSampleWorkspace",
-    projects: [
-        "./Samples/iOS",
-        "./Samples/macOS",
-        "./Samples/tvOS",
-        "./Samples/visionOS",
-        "./Samples/watchOS",
-    ]
-)
